Added File Parser and Build time performance
File Parser has been built into external module.  This has been added
to common target.

Problem with rebuilding vast majority of the code every time been
resolved.  The script in common was doing a copy of headers each time
which meant that Xcode rebuilt common when it didn’t need to.  Given
common was a dependancy for pretty much every module - this meant a
major rebuild.

The header script basically checks if there is a file, removes it (to
resolve previous bug) and creates a sym link.